feat: Wave Link card theming, app theme and per-device apps row (#34)
## Summary
- New "Applications" chips on each Home card: app icon + per-session
peak meter + tooltip with display name and path. Drag a chip to another
render card to reroute that app's output; capture cards show the no-drop
cursor.
- Chips rule-locked when an `ApplicationOutput` rule pins the session,
with a small padlock badge and tooltip naming the rule.
- High-quality icons via `IShellItemImageFactory` at 128px with
premultiplied alpha and a software box-filter downscale to 64 for clean
compositing at any DPI.

## Robustness
- Per-session meters survive session recycle (Expired controls skipped
during dedup so the live control wins).
- `SessionWatcher` subscribes to existing-session state events on attach
so pause/resume/seek transitions actually fire `SessionsChanged` for
long-running apps.
- `SyncCardApps` defends against transient empty enumerations so a COM
hiccup during state churn doesn't mass-wipe chips.
- `SessionsChanged` no longer triggers a full card rebuild - mpv seek
spams these and was tearing down every `ItemsRepeater` visual on every
event. In-place chip reconcile, debounced 200ms.
- Chip revival is strictly peak-based so silent-but-Active sessions get
pruned at 30s and stay pruned.

## Filtering
- System Sounds, Earmark itself, and audio-forwarder processes (Wave
Link, Voicemeeter, VB-Cable, Nvidia Broadcast) dropped from the apps
row.
- Same `System Sounds` filter applied on the Sessions page where it was
duplicating per device.

## App theme
- New Theme setting (System / Light / Dark) under Settings > Appearance,
persisted as a string.
- Mica is applied via `MicaController` (not `<Window.SystemBackdrop>`)
so its tint, the content theme and caption-button colours all follow the
setting; System mode still tracks live OS theme changes.

## Wave Link channel theming
- Optional per-card theming via a new Wave Link > Channel style setting
(Off / Colours / Icons). Colours tints the mute tile with the channel's
accent (the dominant opaque pixel of the channel artwork); Icons swaps
in the channel's own Wave Link bitmap.
- Reads `getChannels` artwork + the mix icon name over the WebSocket;
accents and bitmaps are cached so the 5s snapshot poll doesn't
re-decode. A style change restyles cards in place (no grid rebuild).

## UI
- Shared `ExpanderPill` control backs both the Devices first-rule chip
and the Rules row header so the two expanders look and behave
identically; replaces the per-page toggle handlers.
- Right-click menus: hide/show a device on Devices; enable/disable and
delete a rule on Rules.
- Spacing and corner radii normalised onto the Fluent 2 tokens
(`SpacingXXSmall`..`SpacingXXLarge`, `CardCornerRadius`); Settings
grouped into General / Appearance / System tray / Wave Link /
Diagnostics sections.

## Also in this PR
- Device-card icons matched to the Wave Link mix a device backs
(speakers / headphones / chat / etc.).
- Wave Link input mute and volume routed over the WebSocket (`setMix` /
`setInputDevice`).
- New setting to reconcile Wave Link device names back to their friendly
names.
- Reuse the cached `MMDevice` and skip no-op mute notifications to cut
audio glitches.
- CoreAudio COM kept off the UI thread to stop the periodic-timer hang.

## Test plan
- [ ] mpv seek/pause: chip stays, peak keeps updating, no UI hang
- [ ] Edge / Discord chips render crisp icons matching their taskbar
appearance
- [ ] Drag chip onto another render card; reroute applies
- [ ] Drag chip onto a capture card: cursor shows the no-drop indicator
- [ ] Rule-locked chips show padlock and refuse drag
- [ ] System Sounds and Earmark itself don't appear in chips or Sessions
page
- [ ] App closed: chip disappears within a tick; silent 30s: chip prunes
and stays pruned until audio resumes
- [ ] Theme switch (Light/Dark/System): Mica tint, content and caption
buttons all follow
- [ ] Channel style Colours: mute tiles take the channel accent; Icons:
tiles show the Wave Link bitmap; Off: plain Fluent
- [ ] Right-click a device to hide/show; right-click a rule to
enable/disable and delete
- [ ] Devices first-rule chip and Rules row header expand on a single
click

---

<!-- release-please footers: one bare conventional-commit line per extra
change, separated by blank lines. Title already covers the apps row, app
theme and Wave Link card theming. -->

feat(home): match device-card icons to Wave Link mixes

feat(wavelink): route Wave Link input mute and volume over the WebSocket

feat: add a setting to reconcile Wave Link device names

feat(ui): right-click menus to hide a device and enable or delete a rule

perf(audio): reuse the cached MMDevice and skip no-op mute notifications

fix(audio): keep CoreAudio COM off the UI thread to stop hangs

---------

+++ src/Earmark.Audio/WaveLink/WaveLinkClient.cs
@@ -12,13 +12,14 @@ public sealed class WaveLinkClient : IAsyncDisposable
     private const string OriginHeader = "streamdeck://";
 
     private readonly ILogger<WaveLinkClient> _logger;
-    private readonly ConcurrentDictionary<int, TaskCompletionSource<JsonElement>> _pending = new();
+    private readonly ConcurrentDictionary<long, TaskCompletionSource<JsonElement>> _pending = new();
     private readonly SemaphoreSlim _sendLock = new(1, 1);
     private readonly CancellationTokenSource _disposalCts = new();
 
     private ClientWebSocket? _socket;
     private Task? _receiveLoop;
-    private int _nextId;
+    // long so the per-call counter cannot wrap into the id<=0 range that DispatchFrame ignores.
+    private long _nextId;
 
     public WaveLinkClient(ILogger<WaveLinkClient> logger)
     {
@@ -206,7 +207,7 @@ public sealed class WaveLinkClient : IAsyncDisposable
         try
         {
             var root = doc.RootElement;
-            if (root.TryGetProperty("id", out var idEl) && idEl.ValueKind == JsonValueKind.Number && idEl.TryGetInt32(out var id) && id > 0)
+            if (root.TryGetProperty("id", out var idEl) && idEl.ValueKind == JsonValueKind.Number && idEl.TryGetInt64(out var id) && id > 0)
             {
                 if (_pending.TryRemove(id, out var tcs))
                 {
